+++ app/src/screenshotTestFullDebug/reference/io/homeassistant/companion/android/matter/MatterCommissioningScreenshotTest/MatterCommissioningScreen steps_foldable_c908f502_05386160_6.png
