+++ app/src/screenshotTestFullDebug/reference/io/homeassistant/companion/android/matter/MatterCommissioningScreenshotTest/MatterCommissioningScreen steps_phone_landscape_9e00b29d_05386160_3.png
