+++ app/src/screenshotTestFullDebug/reference/io/homeassistant/companion/android/matter/MatterCommissioningScreenshotTest/MatterCommissioningScreen steps_small_phone_66e7bbf2_05386160_3.png
