+++ app/src/screenshotTestFullDebug/reference/io/homeassistant/companion/android/matter/MatterCommissioningScreenshotTest/MatterCommissioningScreen steps_tablet_2f22c4ea_05386160_5.png
