+++ app/src/screenshotTestFullDebug/reference/io/homeassistant/companion/android/matter/MatterCommissioningScreenshotTest/MatterCommissioningScreen steps_tablet_landscape_62cae397_05386160_4.png
